refactor everything into string-cases
